requirements/test.txt
# production and test requirements

-r prod.txt

pytest>=4.6 # required for pytest-cov
pytest-django
pytest-cov
wheel
# fix issue with taggit fields: https://github.com/makecodes/django-dbml/pull/4
django-dbml
#git+https://github.com/thatbudakguy/django-dbml@patch-1#egg=django-dbml
percy-selenium